+++ src/QuickMarcEditor/utils.test.js
@@ -1994,11 +1994,35 @@ describe('QuickMarcEditor utils', () => {
     });
   });
 
+  describe('validate852OnlyOneSubfield', () => {
+    it ('should return true for only one or none location subfield', () => {
+      expect(utils.validateSingleNoneSubfield({ content: '$b VA/LI/D'}, '$b')).toBe(true);
+    });
+
+    it ('should return true for none location subfield', () => {
+      expect(utils.validateSingleNoneSubfield({ content: '$a VA/LI/D'}, '$b')).toBe(true);
+    });
+
+    it ('sholud retun false for more the one location subfield', () => {
+      expect(utils.validateSingleNoneSubfield({ content: '$b VA/LI/D $b NO/VA/LI/D'}, '$b')).toBe(false);
+    });
+  });
+
+  describe('countSubField', () => {
+    it ('shoud return correct numer of subfield in content of field', () => {
+      expect(utils.countSubField({ content: '$b   $bb$b '}, '$b')).toEqual(3);
+    });
+  });
+
   describe('validateLocationSubfield', () => {
     it('should return true for valid location subfield', () => {
       expect(utils.validateLocationSubfield({ content: '$b VA/LI/D ' }, locations)).toBe(true);
     });
 
+    it('should return true for valid location subfield even more space between', () => {
+      expect(utils.validateLocationSubfield({ content: '$b    VA/LI/D ' }, locations)).toBe(true);
+    });
+    
     it('should return false for locations that do not exist', () => {
       expect(utils.validateLocationSubfield({ content: '$b NOT/VA/LI/D ' }, locations)).toBe(false);
     });